refactor: remove exception subpackage
