+++ lib/src/main/java/hu/bme/mit/ftsrg/hypernate/HypernateException.java
@@ -1,5 +1,5 @@
 /* SPDX-License-Identifier: Apache-2.0 */
-package hu.bme.mit.ftsrg.hypernate.exception;
+package hu.bme.mit.ftsrg.hypernate;
 
 import lombok.experimental.StandardException;
