+++ lib/src/main/java/hu/bme/mit/ftsrg/hypernate/entity/DataAccessException.java
@@ -1,7 +1,7 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 package hu.bme.mit.ftsrg.hypernate.entity;
 
-import hu.bme.mit.ftsrg.hypernate.exception.HypernateException;
+import hu.bme.mit.ftsrg.hypernate.HypernateException;
 import lombok.experimental.StandardException;
 
 /** Base class for {@link HypernateException}s related to data access problems. */
